Add neighbour nodes

+++ Utilities/trimesh2graph.m
@@ -26,12 +26,12 @@ for ii=1:size(TRI,1)
     u = unique(elem);
     
     
-    graph.elements(ii).neighbours = sort(u(histc(elem,u)==nD));
+    graph.elements(ii).neighbours = uint32(sort(u(histc(elem,u)==nD)));
     
     graph.elements(ii).nodeId=TRI(ii,:);
     
     for jj=1:nD+1
-        graph.nodes(TRI(ii,jj)).neighbour_elems=[graph.nodes(TRI(ii,jj)).neighbour_elems,ii];
+        graph.nodes(TRI(ii,jj)).neighbour_elems=uint32([graph.nodes(TRI(ii,jj)).neighbour_elems,ii]);
     end
     
     % The amount of neihgbours must be nd+1 or nd (in case of the boundary)
@@ -40,7 +40,7 @@ end
 
 
 for ii=1:length(graph.nodes)
-     graph.nodes(ii).neighbour_elems=sort( graph.nodes(ii).neighbour_elems);
+     graph.nodes(ii).neighbour_elems=sort(graph.nodes(ii).neighbour_elems);
 end
 
 % Check which of the elements lie in a boundary of the mesh
